pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prediction-rts"
version = "0.1.0"
description = "RTS Agent Behavior Prediction and Decision Making"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "PredictionRTS Team"}
]
keywords = ["reinforcement-learning", "transformer", "starcraft2", "rts", "decision-transformer"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=1.10.0",
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "scipy>=1.7.0",
    "scikit-learn>=0.24.0",
    "matplotlib>=3.4.0",
    "seaborn>=0.11.0",
    "tqdm>=4.62.0",
    "omegaconf>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "black>=21.0",
    "isort>=5.0",
    "flake8>=3.9",
    "mypy>=0.900",
]
sc2 = [
    "pysc2>=3.0.0",
    "absl-py>=0.13.0",
]
live = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "requests>=2.28.0",
    "sse-starlette>=1.6.0",
    "streamlit>=1.28.0",
    "plotly>=5.0.0",
    "pyvis>=0.3.0",
]
web = [
    "streamlit>=1.28.0",
    "plotly>=5.0.0",
    "pyvis>=0.3.0",
    "requests>=2.28.0",
]

[project.scripts]
predict-rtstrain = "scripts.run_train:main"
predict-rts-evaluate = "scripts.run_evaluate:main"
predict-rts-predict = "scripts.run_predict:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "configs*"]

[tool.black]
line-length = 100
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
